Linkedin_Scraper/requirements.txt
numpy==1.22.2
pandas==1.4.1
python-dotenv==0.20.0
selenium==4.1.5
pyinstaller==5.0.1